Trim ITs down to one per dataset, add dataset to test name factor out some common test utility code
