+++ test/integration_tests/test_print_files.py
@@ -4,39 +4,13 @@ from pathlib import Path
 from time import sleep
 
 import paramiko
-import pgpy
 import pytest
 
 from app.constants import ActionType, PackCode
 from config import TestConfig
 from test.integration_tests.utilities import build_test_messages, send_action_messages, ICL_message_template, \
     print_questionnaire_message_template, P_OR_message_template, PPD1_3_message_template, reminder_message_template
-
-
-def test_ICL1E(sftp_client):
-    # Given
-    icl1e_messages, _ = build_test_messages(ICL_message_template, 1, 'ICL1E', 'P_IC_ICL1')
-    send_action_messages(icl1e_messages)
-
-    # When
-    matched_manifest_file, matched_print_file = get_print_and_manifest_filenames(sftp_client,
-                                                                                 TestConfig.SFTP_PPO_DIRECTORY,
-                                                                                 'P_IC_ICL1')
-
-    # Then
-    decrypted_print_file = get_and_check_print_file(
-        sftp=sftp_client,
-        remote_print_file_path=TestConfig.SFTP_PPO_DIRECTORY + matched_print_file,
-        decryption_key_path=Path(__file__).parents[2].joinpath('dummy_keys',
-                                                               'dummy_ppo_supplier_private_key.asc'),
-        decryption_key_passphrase='test',
-        expected='0|test_caseref||||123 Fake Street|Duffryn||Newport|NPXXXX|P_IC_ICL1||||\n')
-
-    get_and_check_manifest_file(sftp=sftp_client,
-                                remote_manifest_path=TestConfig.SFTP_PPO_DIRECTORY + matched_manifest_file,
-                                expected_values={
-                                    'description': 'Initial contact letter households - England',
-                                    'dataset': 'PPD1.1'}, decrypted_print_file=decrypted_print_file)
+from test.utilities import decrypt_message
 
 
 def test_ICL1E_split_files(sftp_client):
@@ -84,39 +58,114 @@ def test_ICL1E_split_files(sftp_client):
                   '20|test_caseref||||123 Fake Street|Duffryn||Newport|NPXXXX|P_IC_ICL1||||\n'))
 
 
-def test_ICHHQE(sftp_client):
+def test_ICL1E_PPD1_1(sftp_client):
     # Given
-    ichhqe_messages, _ = build_test_messages(print_questionnaire_message_template, 3, 'ICHHQE', 'P_IC_H1')
-    send_action_messages(ichhqe_messages)
+    icl1e_messages, _ = build_test_messages(ICL_message_template, 1, 'ICL1E', 'P_IC_ICL1')
+    send_action_messages(icl1e_messages)
 
     # When
     matched_manifest_file, matched_print_file = get_print_and_manifest_filenames(sftp_client,
-                                                                                 TestConfig.SFTP_QM_DIRECTORY,
-                                                                                 'P_IC_H1')
+                                                                                 TestConfig.SFTP_PPO_DIRECTORY,
+                                                                                 'P_IC_ICL1')
 
     # Then
     decrypted_print_file = get_and_check_print_file(
         sftp=sftp_client,
-        remote_print_file_path=TestConfig.SFTP_QM_DIRECTORY + matched_print_file,
+        remote_print_file_path=TestConfig.SFTP_PPO_DIRECTORY + matched_print_file,
         decryption_key_path=Path(__file__).parents[2].joinpath('dummy_keys',
-                                                               'dummy_qm_supplier_private_key.asc'),
-        decryption_key_passphrase='supplier',
-        expected=(
-            '0|english_qid|1|welsh_qid|test_qm_coordinator_id|'
-            '|||123 Fake Street|Duffryn||Newport|NPXXXX|P_IC_H1||\n'
-            '2|english_qid|3|welsh_qid|test_qm_coordinator_id|'
-            '|||123 Fake Street|Duffryn||Newport|NPXXXX|P_IC_H1||\n'
-            '4|english_qid|5|welsh_qid|test_qm_coordinator_id|'
-            '|||123 Fake Street|Duffryn||Newport|NPXXXX|P_IC_H1||\n'))
+                                                               'dummy_ppo_supplier_private_key.asc'),
+        decryption_key_passphrase='test',
+        expected='0|test_caseref||||123 Fake Street|Duffryn||Newport|NPXXXX|P_IC_ICL1||||\n')
 
     get_and_check_manifest_file(sftp=sftp_client,
-                                remote_manifest_path=TestConfig.SFTP_QM_DIRECTORY + matched_manifest_file,
+                                remote_manifest_path=TestConfig.SFTP_PPO_DIRECTORY + matched_manifest_file,
                                 expected_values={
-                                    'description': 'Initial contact questionnaire households - England',
-                                    'dataset': 'QM3.2'}, decrypted_print_file=decrypted_print_file)
+                                    'description': 'Initial contact letter households - England',
+                                    'dataset': 'PPD1.1'}, decrypted_print_file=decrypted_print_file)
 
 
-def test_ICHHQW(sftp_client):
+def test_P_RL_1RL1A_PPD1_2(sftp_client):
+    # Given
+    messages, _ = build_test_messages(reminder_message_template, 1, 'P_RL_1RL1A', 'P_RL_1RL1A')
+    send_action_messages(messages)
+
+    # When
+    matched_manifest_file, matched_print_file = get_print_and_manifest_filenames(sftp_client,
+                                                                                 TestConfig.SFTP_PPO_DIRECTORY,
+                                                                                 'P_RL_1RL1A')
+
+    # Then
+    decrypted_print_file = get_and_check_print_file(
+        sftp=sftp_client,
+        remote_print_file_path=TestConfig.SFTP_PPO_DIRECTORY + matched_print_file,
+        decryption_key_path=Path(__file__).parents[2].joinpath('dummy_keys',
+                                                               'dummy_ppo_supplier_private_key.asc'),
+        decryption_key_passphrase='test',
+        expected='0|test_caseref||||123 Fake Street|Duffryn||Newport|NPXXXX|P_RL_1RL1A||||\n')
+
+    get_and_check_manifest_file(sftp=sftp_client,
+                                remote_manifest_path=TestConfig.SFTP_PPO_DIRECTORY + matched_manifest_file,
+                                expected_values={
+                                    'description':
+                                        '1st Reminder, Letter - for England addresses'
+                                        ' for survey launched but not completed',
+                                    'dataset': 'PPD1.2'}, decrypted_print_file=decrypted_print_file)
+
+
+def test_P_LP_HL1_PPD1_3(sftp_client):
+    # Given
+    messages, _ = build_test_messages(PPD1_3_message_template, 1, 'P_LP_HLX', 'P_LP_HL1', uac=False)
+    send_action_messages(messages)
+
+    # When
+    matched_manifest_file, matched_print_file = get_print_and_manifest_filenames(sftp_client,
+                                                                                 TestConfig.SFTP_PPO_DIRECTORY,
+                                                                                 'P_LP_HL1')
+
+    # Then
+    decrypted_print_file = get_and_check_print_file(
+        sftp=sftp_client,
+        remote_print_file_path=TestConfig.SFTP_PPO_DIRECTORY + matched_print_file,
+        decryption_key_path=Path(__file__).parents[2].joinpath('dummy_keys',
+                                                               'dummy_ppo_supplier_private_key.asc'),
+        decryption_key_passphrase='test',
+        expected='|test_caseref|Mr|Test|McTest|123 Fake Street|Duffryn||Newport|NPXXXX|P_LP_HL1||||\n')
+
+    get_and_check_manifest_file(sftp=sftp_client,
+                                remote_manifest_path=TestConfig.SFTP_PPO_DIRECTORY + matched_manifest_file,
+                                expected_values={
+                                    'description': 'Household Questionnaire Large Print pack for England',
+                                    'dataset': 'PPD1.3'}, decrypted_print_file=decrypted_print_file)
+
+
+def test_CE1_IC01_PPD1_7(sftp_client):
+    # Given
+    ce1_ic01_messages, _ = build_test_messages(ICL_message_template, 1, 'CE1_IC01', 'D_CE1A_ICLCR1', ce=True)
+    send_action_messages(ce1_ic01_messages)
+
+    # When
+    matched_manifest_file, matched_print_file = get_print_and_manifest_filenames(sftp_client,
+                                                                                 TestConfig.SFTP_PPO_DIRECTORY,
+                                                                                 'D_CE1A_ICLCR1')
+
+    # Then
+    decrypted_print_file = get_and_check_print_file(
+        sftp=sftp_client,
+        remote_print_file_path=TestConfig.SFTP_PPO_DIRECTORY + matched_print_file,
+        decryption_key_path=Path(__file__).parents[2].joinpath('dummy_keys',
+                                                               'dummy_ppo_supplier_private_key.asc'),
+        decryption_key_passphrase='test',
+        expected='0|test_caseref||||123 Fake Street|Duffryn||Newport|NPXXXX|D_CE1A_ICLCR1|english_qid|ONS'
+                 '|ppo_field_coordinator_id|dummy_field_officer_id\n')
+
+    get_and_check_manifest_file(sftp=sftp_client,
+                                remote_manifest_path=TestConfig.SFTP_PPO_DIRECTORY + matched_manifest_file,
+                                expected_values={
+                                    'description': 'CE1 ICL with UAC for England (Hand Delivery) Addressed',
+                                    'dataset': 'PPD1.7'}, decrypted_print_file=decrypted_print_file)
+
+
+def test_ICHHQW_QM3_2(sftp_client):
     # Given
     ichhqw_messages, _ = build_test_messages(print_questionnaire_message_template, 3, 'ICHHQW', 'P_IC_H2')
     send_action_messages(ichhqw_messages)
@@ -148,117 +197,7 @@ def test_ICHHQW(sftp_client):
                                     'dataset': 'QM3.2'}, decrypted_print_file=decrypted_print_file)
 
 
-def test_P_OR_H2W(sftp_client):
-    # Given
-    messages, _ = build_test_messages(P_OR_message_template, 3, 'P_OR_HX', 'P_OR_H2W')
-    send_action_messages(messages)
-
-    # When
-    matched_manifest_file, matched_print_file = get_print_and_manifest_filenames(sftp_client,
-                                                                                 TestConfig.SFTP_QM_DIRECTORY,
-                                                                                 'P_OR_H2W')
-
-    # Then
-    decrypted_print_file = get_and_check_print_file(
-        sftp=sftp_client,
-        remote_print_file_path=TestConfig.SFTP_QM_DIRECTORY + matched_print_file,
-        decryption_key_path=Path(__file__).parents[2].joinpath('dummy_keys',
-                                                               'dummy_qm_supplier_private_key.asc'),
-        decryption_key_passphrase='supplier',
-        expected=(
-            '0|english_qid||||Mr|Test|McTest|123 Fake Street|Duffryn||Newport|NPXXXX|P_OR_H2W||\n'
-            '1|english_qid||||Mr|Test|McTest|123 Fake Street|Duffryn||Newport|NPXXXX|P_OR_H2W||\n'
-            '2|english_qid||||Mr|Test|McTest|123 Fake Street|Duffryn||Newport|NPXXXX|P_OR_H2W||\n'))
-
-    get_and_check_manifest_file(sftp=sftp_client,
-                                remote_manifest_path=TestConfig.SFTP_QM_DIRECTORY + matched_manifest_file,
-                                expected_values={
-                                    'description': 'Household Questionnaire for Wales (Welsh)',
-                                    'dataset': 'QM3.4'}, decrypted_print_file=decrypted_print_file)
-
-
-def test_P_RL_1RL1A(sftp_client):
-    # Given
-    messages, _ = build_test_messages(reminder_message_template, 1, 'P_RL_1RL1A', 'P_RL_1RL1A')
-    send_action_messages(messages)
-
-    # When
-    matched_manifest_file, matched_print_file = get_print_and_manifest_filenames(sftp_client,
-                                                                                 TestConfig.SFTP_PPO_DIRECTORY,
-                                                                                 'P_RL_1RL1A')
-
-    # Then
-    decrypted_print_file = get_and_check_print_file(
-        sftp=sftp_client,
-        remote_print_file_path=TestConfig.SFTP_PPO_DIRECTORY + matched_print_file,
-        decryption_key_path=Path(__file__).parents[2].joinpath('dummy_keys',
-                                                               'dummy_ppo_supplier_private_key.asc'),
-        decryption_key_passphrase='test',
-        expected='0|test_caseref||||123 Fake Street|Duffryn||Newport|NPXXXX|P_RL_1RL1A||||\n')
-
-    get_and_check_manifest_file(sftp=sftp_client,
-                                remote_manifest_path=TestConfig.SFTP_PPO_DIRECTORY + matched_manifest_file,
-                                expected_values={
-                                    'description':
-                                        '1st Reminder, Letter - for England addresses'
-                                        ' for survey launched but not completed',
-                                    'dataset': 'PPD1.2'}, decrypted_print_file=decrypted_print_file)
-
-
-def test_P_LP_HL1(sftp_client):
-    # Given
-    messages, _ = build_test_messages(PPD1_3_message_template, 1, 'P_LP_HLX', 'P_LP_HL1', uac=False)
-    send_action_messages(messages)
-
-    # When
-    matched_manifest_file, matched_print_file = get_print_and_manifest_filenames(sftp_client,
-                                                                                 TestConfig.SFTP_PPO_DIRECTORY,
-                                                                                 'P_LP_HL1')
-
-    # Then
-    decrypted_print_file = get_and_check_print_file(
-        sftp=sftp_client,
-        remote_print_file_path=TestConfig.SFTP_PPO_DIRECTORY + matched_print_file,
-        decryption_key_path=Path(__file__).parents[2].joinpath('dummy_keys',
-                                                               'dummy_ppo_supplier_private_key.asc'),
-        decryption_key_passphrase='test',
-        expected='|test_caseref|Mr|Test|McTest|123 Fake Street|Duffryn||Newport|NPXXXX|P_LP_HL1||||\n')
-
-    get_and_check_manifest_file(sftp=sftp_client,
-                                remote_manifest_path=TestConfig.SFTP_PPO_DIRECTORY + matched_manifest_file,
-                                expected_values={
-                                    'description': 'Household Questionnaire Large Print pack for England',
-                                    'dataset': 'PPD1.3'}, decrypted_print_file=decrypted_print_file)
-
-
-def test_P_OR_I2(sftp_client):
-    # Given
-    messages, _ = build_test_messages(print_questionnaire_message_template, 1, 'P_OR_IX', 'P_OR_I2')
-    send_action_messages(messages)
-
-    # When
-    matched_manifest_file, matched_print_file = get_print_and_manifest_filenames(sftp_client,
-                                                                                 TestConfig.SFTP_QM_DIRECTORY,
-                                                                                 'P_OR_I2')
-    # Then
-    decrypted_print_file = get_and_check_print_file(
-        sftp=sftp_client,
-        remote_print_file_path=TestConfig.SFTP_QM_DIRECTORY + matched_print_file,
-        decryption_key_path=Path(__file__).parents[2].joinpath('dummy_keys',
-                                                               'dummy_qm_supplier_private_key.asc'),
-        decryption_key_passphrase='supplier',
-        expected=(
-            '0|english_qid|1|welsh_qid|test_qm_coordinator_id|'
-            '|||123 Fake Street|Duffryn||Newport|NPXXXX|P_OR_I2||\n'))
-
-    get_and_check_manifest_file(sftp=sftp_client,
-                                remote_manifest_path=TestConfig.SFTP_QM_DIRECTORY + matched_manifest_file,
-                                expected_values={
-                                    'description': 'Individual Questionnaire for Wales (in English)',
-                                    'dataset': 'QM3.4'}, decrypted_print_file=decrypted_print_file)
-
-
-def test_P_QU_H2(sftp_client):
+def test_P_QU_H2_QM3_3(sftp_client):
     # Given
     messages, _ = build_test_messages(print_questionnaire_message_template, 3, ActionType.P_QU_H2.value,
                                       PackCode.P_QU_H2.value)
@@ -291,83 +230,33 @@ def test_P_QU_H2(sftp_client):
                                     'dataset': 'QM3.3'}, decrypted_print_file=decrypted_print_file)
 
 
-def test_P_RD_2RL2B_1(sftp_client):
+def test_P_OR_H2W_QM3_4(sftp_client):
     # Given
-    messages, _ = build_test_messages(reminder_message_template, 1, 'P_RD_2RL2B_1', 'P_RD_2RL2B_1')
+    messages, _ = build_test_messages(P_OR_message_template, 3, 'P_OR_HX', 'P_OR_H2W')
     send_action_messages(messages)
 
     # When
     matched_manifest_file, matched_print_file = get_print_and_manifest_filenames(sftp_client,
-                                                                                 TestConfig.SFTP_PPO_DIRECTORY,
-                                                                                 'P_RD_2RL2B_1')
+                                                                                 TestConfig.SFTP_QM_DIRECTORY,
+                                                                                 'P_OR_H2W')
 
     # Then
     decrypted_print_file = get_and_check_print_file(
         sftp=sftp_client,
-        remote_print_file_path=TestConfig.SFTP_PPO_DIRECTORY + matched_print_file,
+        remote_print_file_path=TestConfig.SFTP_QM_DIRECTORY + matched_print_file,
         decryption_key_path=Path(__file__).parents[2].joinpath('dummy_keys',
-                                                               'dummy_ppo_supplier_private_key.asc'),
-        decryption_key_passphrase='test',
-        expected='0|test_caseref||||123 Fake Street|Duffryn||Newport|NPXXXX|P_RD_2RL2B_1||||\n')
+                                                               'dummy_qm_supplier_private_key.asc'),
+        decryption_key_passphrase='supplier',
+        expected=(
+            '0|english_qid||||Mr|Test|McTest|123 Fake Street|Duffryn||Newport|NPXXXX|P_OR_H2W||\n'
+            '1|english_qid||||Mr|Test|McTest|123 Fake Street|Duffryn||Newport|NPXXXX|P_OR_H2W||\n'
+            '2|english_qid||||Mr|Test|McTest|123 Fake Street|Duffryn||Newport|NPXXXX|P_OR_H2W||\n'))
 
     get_and_check_manifest_file(sftp=sftp_client,
-                                remote_manifest_path=TestConfig.SFTP_PPO_DIRECTORY + matched_manifest_file,
+                                remote_manifest_path=TestConfig.SFTP_QM_DIRECTORY + matched_manifest_file,
                                 expected_values={
-                                    'description': 'Response driven reminder group 1 Welsh',
-                                    'dataset': 'PPD1.2'}, decrypted_print_file=decrypted_print_file)
-
-
-def test_CE1_IC01(sftp_client):
-    # Given
-    ce1_ic01_messages, _ = build_test_messages(ICL_message_template, 1, 'CE1_IC01', 'D_CE1A_ICLCR1', ce=True)
-    send_action_messages(ce1_ic01_messages)
-
-    # When
-    matched_manifest_file, matched_print_file = get_print_and_manifest_filenames(sftp_client,
-                                                                                 TestConfig.SFTP_PPO_DIRECTORY,
-                                                                                 'D_CE1A_ICLCR1')
-
-    # Then
-    decrypted_print_file = get_and_check_print_file(
-        sftp=sftp_client,
-        remote_print_file_path=TestConfig.SFTP_PPO_DIRECTORY + matched_print_file,
-        decryption_key_path=Path(__file__).parents[2].joinpath('dummy_keys',
-                                                               'dummy_ppo_supplier_private_key.asc'),
-        decryption_key_passphrase='test',
-        expected='0|test_caseref||||123 Fake Street|Duffryn||Newport|NPXXXX|D_CE1A_ICLCR1|english_qid|ONS'
-                 '|ppo_field_coordinator_id|dummy_field_officer_id\n')
-
-    get_and_check_manifest_file(sftp=sftp_client,
-                                remote_manifest_path=TestConfig.SFTP_PPO_DIRECTORY + matched_manifest_file,
-                                expected_values={
-                                    'description': 'CE1 ICL with UAC for England (Hand Delivery) Addressed',
-                                    'dataset': 'PPD1.7'}, decrypted_print_file=decrypted_print_file)
-
-
-def test_SPG_IC11(sftp_client):
-    # Given
-    spg_ic11_messages, _ = build_test_messages(ICL_message_template, 1, 'SPG_IC11', 'P_ICCE_ICL1')
-    send_action_messages(spg_ic11_messages)
-
-    # When
-    matched_manifest_file, matched_print_file = get_print_and_manifest_filenames(sftp_client,
-                                                                                 TestConfig.SFTP_PPO_DIRECTORY,
-                                                                                 'P_ICCE_ICL1')
-
-    # Then
-    decrypted_print_file = get_and_check_print_file(
-        sftp=sftp_client,
-        remote_print_file_path=TestConfig.SFTP_PPO_DIRECTORY + matched_print_file,
-        decryption_key_path=Path(__file__).parents[2].joinpath('dummy_keys',
-                                                               'dummy_ppo_supplier_private_key.asc'),
-        decryption_key_passphrase='test',
-        expected='0|test_caseref||||123 Fake Street|Duffryn||Newport|NPXXXX|P_ICCE_ICL1||||\n')
-
-    get_and_check_manifest_file(sftp=sftp_client,
-                                remote_manifest_path=TestConfig.SFTP_PPO_DIRECTORY + matched_manifest_file,
-                                expected_values={
-                                    'description': 'Household ICL with UAC for England (Post Out) Addressed',
-                                    'dataset': 'PPD1.7'}, decrypted_print_file=decrypted_print_file)
+                                    'description': 'Household Questionnaire for Wales (Welsh)',
+                                    'dataset': 'QM3.4'}, decrypted_print_file=decrypted_print_file)
 
 
 def test_our_decryption_key(sftp_client):
@@ -469,14 +358,6 @@ def get_and_check_multiple_print_files(sftp, remote_print_file_directory, print_
                                                    decryption_key_passphrase)
         assert decrypted_print_file in expected
         assert len(decrypted_print_file.splitlines()) == 10
-
-
-def decrypt_message(message, key_file_path, key_passphrase):
-    key, _ = pgpy.PGPKey.from_file(key_file_path)
-    with key.unlock(key_passphrase):
-        encrypted_text_message = pgpy.PGPMessage.from_blob(message)
-        message_text = key.decrypt(encrypted_text_message)
-        return message_text.message
 
 
 @pytest.fixture(autouse=True)
