+++ test/unit_tests/app/test_encryption.py
@@ -1,10 +1,10 @@
 from pathlib import Path
 
-import pgpy
 import pytest
 
 from app.constants import Supplier
 from app.encryption import pgp_encrypt_message
+from test.utilities import decrypt_message
 
 DUMMY_KEYS_PATH = Path(__file__).parents[3].joinpath('dummy_keys')
 
@@ -30,11 +30,3 @@ def test_pgp_encrypt_message(supplier, supplier_key_path, supplier_passphrase):
 
     assert supplier_decrypted_message == message, 'Supplier key should be able to decrypt the message'
     assert rm_decrypted_message == message, 'RM key should be able to decrypt the message'
-
-
-def decrypt_message(message, key_file_path, key_passphrase):
-    key, _ = pgpy.PGPKey.from_file(key_file_path)
-    with key.unlock(key_passphrase):
-        encrypted_text_message = pgpy.PGPMessage.from_blob(message)
-        message_text = key.decrypt(encrypted_text_message)
-        return message_text.message
